services/agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.0
greenlet==3.0.3
# AI agent
streamlit>=1.32.0
langgraph>=0.1.0
langchain>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
# OTel
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
